feat: split table and root
